requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
dotenv==0.9.9
eval_type_backport==0.3.1
flatbuffers==25.12.19
frozenlist==1.8.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-speech==2.38.0
google-cloud-texttospeech==2.36.0
google-genai==1.73.1
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.3.0
langchain-ollama==1.1.0
langchain-openai==1.2.0
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.34
livekit==1.1.5
livekit-agents==1.5.6
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-deepgram==1.5.6
livekit-plugins-google==1.5.6
livekit-plugins-openai==1.5.6
livekit-plugins-silero==1.5.6
livekit-protocol==1.1.6
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.4
ollama==0.6.1
onnxruntime==1.25.0
openai==2.32.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pillow==12.2.0
pinecone==8.1.2
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
prometheus_client==0.25.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.2
types-protobuf==7.34.1.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
